Pattern based verb arguments
Here are some examples of verbs which can be configured now.

	[[verbs]]
	invocation = "mkdir {subpath}"
	execution = "/bin/mkdir -p {directory}/{subpath}"

As before, when only one verb possibly matches, you don't have to type its
complete name. You can do

	mk new folder

(and yes, space are ok)

	[[verbs]]
	invocation = "create {subpath}"
	execution = "/usr/bin/nvim {directory}/{subpath}"

	[[verbs]]
	invocation = "rm"
	execution = "/bin/rm -rf {file}"
	confirm = true
	leave_broot = false

	[[verbs]]
	invocation = "mv {subpath}"
	execution = "/bin/mv {file} {parent}/{subpath}"

Here's a more complex one, creating a directory then a file in neovim, by
destructuring the argument in name and suffix:

	[[verbs]]
	invocation = "blop {name}\\.{type}"
	execution = "/bin/mkdir {parent}/{type} && /usr/bin/nvim {parent}/{type}/{name}.{type}"
	from_shell = true

Notice the `\\.` ? That's because the invocation pattern is interpreted as a regular expression
(with just a shortcut for the easy case, enablinrg `{name}`).
The whole regular expression syntax may be useful for more complex rules.
Let's say we don't want the type to contain dots, then we do this:

	[[verbs]]
	invocation = "blop {name}\\.(?P<type>[^.]+)"
	execution = "/bin/mkdir {parent}/{type} && /usr/bin/nvim {parent}/{type}/{name}.{type}"
	from_shell = true