smart path normalization
Given a verb like this one:

	[[verbs]]
	invocation = "create {subpath}"
	execution = "/usr/bin/nvim {directory}/{subpath}"

it's possible for example to type an argument starting with ../../ and
thus get a non normalized path which isn't pretty on screen.

When broot recognizes a path, and the file/dir exists, it normalizes it.

Bonus: this makes it clear when the typed subpath produces a path to an
existing file/dir (which matters for example when the command is `mv`.