faster fuzzy searches
The `find` function do a lot of computations whose purpose
is to find the exact position in chars of the best match.
I've made another function, `score_of`, which only returns
the score of the best match. This function is about 3 times
faster.
In order to manage the code duplication this brought, I've
added a few test units.
The code has also been modified to be benchmarkable and a
bench case measures fuzzy searches.