internals can accept an argument. :focus does accept one
Both in input and in configuration.
This allowed to remove the :focus_user_home and :focus_root
internals.

It's now possible to define a verb like

```
[[verbs]]
key = "ctrl-h"
invocation = "home"
execution = ":focus ~/dev"
```

which can be called as ctrl-h, :home, but also :!home
(in order to open the ~/dev directory in a new panel)