compute pending tasks in non focused panels too
There are a few cases where it's useful.

One of them is when you have an unfocused panel displaying the sizes and
you do an operation changing the sizes (like copying a target
directory). In such a case you want to be notified that the sizes
changed (like your project being 99GB heavier).