improves and simplifies fuzzy matching (WIP)
I've removed the separation between a scoring algorithm and
the matching one. This means we lose a few % in performances
but it's way easier to change the matching algorighm when it
doesn't have to be reproduced with a lighter logic.

I've tuned the algorithm to avoid singled chars if possible
and to have a malus when they can't be avoided.