:write_output & :clear_output, --verb_output (#834)
Fix: #825

Add a `--verb-output` launch argument which takes a path to a file (which will be created if necessary)

Add a  `:write_output` internal which allows adding a line to that file. No escaping is done (contrary to what happens with `--outcmd`).

Add a `:clear_output` internal which clears the file.

Here are 2 examples of verbs:

```
    {
        invocation: "wc {cmd}"
        execution: ":write_output wc:{cmd} {file-stem}.bro"
    }

    {
        key: alt-w
        cmd: ":clear_output;:write_output {directory};:quit"
    }
```

The first one is called with an input like `:wc hop` which appends to the output a line like `wc:hop main.bro`. 

The second one makes the content of the output file the directory closest to the selection then quits. It could for example be used for a new version of the `br` shell function.

Note: Semantics isn't pretty. If you have a better idea than "output", please tell me.