option to reset terminal title on exit
It's done with

    reset_terminal_title_on_exit: false

Fix #966

It may be arged that #966 isn't fixed: the title we reset to is
the default one of the terminal, not the one it had just before
broot was launched.
The reason for this is that modern terminals, for security reasons,
don't let you read (and thus restore) some properties of the terminal,
among them its title.