Fix total search impossible after refresh
A refresh operation occurs on many cases, for example after execution
of an external program (including an editor) because those programs
often change files.
Refresh should be fast, which is why it's never doing a "total search".
The bug is that the tree didn't know it wasn't anymore in total search
mode, so it prevented a total search by pretending it wasn't needed.

Fix #985