Watch (#1041)
Add the `:toggle_watch` internal, with shortcut `:watch` and bound to the `alt-w` key combination.

When the current tree is watched, changes (file or directory creation/deletion/move/write) will lead to a refresh.

When there's a failure to watch (which is likely on huge directories) broot leaves watching mode (but you may reenter it again, preferably on a smaller dir).

Fix #730