fix non-trash platforms
279a4af removed the `trash` feature in favor of just using conditional
compilation for supported platforms; but it didn't actually compile on
platforms not supported by the `trash` crate.

This fixes compilation in these cases. Rather than compile out the
trash command altogether, this prints an error message if trash is not
supported on the platform. This way, key bindings and such can still
work across platforms.